+++ excel_workbook_functions.py
@@ -1,11 +1,7 @@
-from file_name_helpers import get_latest_filename, name_next_sheet, make_file_name
 
 from RPA.Excel.Files import Files
-from RPA.Tables import Tables
 from RPA.FileSystem import FileSystem
 
-
-TABLE_MANIPULATOR = Tables()
 FILE_SYSTEM_CONTROLLER = FileSystem()
 
 # Parameters for modifying spread sheet
@@ -17,19 +13,6 @@ LATE_MONTH_BILLS = 'D'
 BILLS_FOR_PAY_PERIOD = 'F'
 PAY_PERIOD_LENGTH_IN_DAYS = 14
 
-
-# This gets the last pay period workbook and makes a copy of it for the current pay period and returns the filename
-def clone_budget_spreadsheet():
-    file_list_from_directory = FILE_SYSTEM_CONTROLLER.list_files_in_directory(BUDGET_SPREADSHEET_DIRECTORY)
-    date_of_last_pay = get_latest_filename(file_list_from_directory)    
-
-    LAST_SPREADSHEET_NAME = make_file_name(date_of_last_pay)
-    NEW_SPREADSHEET_NAME = name_next_sheet(date_of_last_pay, PAY_PERIOD_LENGTH_IN_DAYS)
-
-    FILE_SYSTEM_CONTROLLER.copy_file(BUDGET_SPREADSHEET_DIRECTORY + '/' + LAST_SPREADSHEET_NAME, BUDGET_SPREADSHEET_DIRECTORY + '/' + NEW_SPREADSHEET_NAME)
-
-    print (NEW_SPREADSHEET_NAME, " has been created.")
-    return NEW_SPREADSHEET_NAME
 
 # This function will copy a column of values from one column to another
 def copy_worksheet_column(workbook ,sheet, source, destination):
